pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gencast-repro"
version = "0.1.0"
description = "Public PyTorch reimplementation of a GenCast-style diffusion weather forecaster"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
  {name = "Codex"},
]
dependencies = [
  "numpy>=1.23",
  "scipy>=1.10",
  "xarray>=2023.1.0",
  "pyyaml>=6.0",
  "torch>=2.1",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
gencast-repro = "gencast_repro.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

